refactor: 온보딩 화면 액티비티로 리팩토링
