+++ app/src/main/java/com/egobook/app/MainActivity.kt
@@ -34,7 +34,6 @@ class MainActivity : AppCompatActivity(), BlurController, NotificationController
         // 목적지 변경 리스너 추가: 특정 프래그먼트에서 바텀바 숨기기
         navController.addOnDestinationChangedListener { _, destination, _ ->
             when (destination.id) {
-                R.id.onboardingContainerFragment, // 온보딩 화면
                 R.id.diaryWriteFragment,        // 일기 작성 화면
                 R.id.calenderFragment, // 달력 화면
                 R.id.storeFragment,
