--- app/src/main/java/com/egobook/app/ui/onboarding/view/OnboardingContainerFragment.kt
@@ -1,49 +1,0 @@
-package com.egobook.app.ui.onboarding.view
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
-import com.egobook.app.databinding.FragmentOnboardingContainerBinding
-import com.egobook.app.ui.onboarding.adapter.OnboardingVPAdapter
-class OnboardingContainerFragment : Fragment() {
-
-    private var _binding: FragmentOnboardingContainerBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentOnboardingContainerBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        // 하단 시스템 바 영역만큼 패딩 주기
-        ViewCompat.setOnApplyWindowInsetsListener(binding.root) { v, insets ->
-            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
-            // 기존 패딩은 유지하면서 하단만 시스템 바 높이만큼 추가
-            v.setPadding(v.paddingLeft, v.paddingTop, v.paddingRight, systemBars.bottom)
-            insets
-        }
-
-        setAdapter()
-    }
-
-    private fun setAdapter() {
-        val onboardingVPAdapter = OnboardingVPAdapter(this)
-        binding.vpOnboarding.adapter = onboardingVPAdapter
-        binding.circleIndicator.setViewPager(binding.vpOnboarding)
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null // 메모리 누수 방지를 위해 필수
-    }
-}
