+++ app/src/main/res/navigation/bottom_navigation.xml
@@ -3,7 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/bottom_navigation"
-    app:startDestination="@id/onboardingContainerFragment">
+    app:startDestination="@id/menu_home">
 
     <fragment
         android:id="@+id/menu_home"
@@ -133,19 +133,6 @@
         <argument app:argType="com.egobook.app.ui.counseling.model.WeeklyReportModel"
             android:name="weeklyReportItem"
             app:nullable="false"/>
-    </fragment>
-
-    <!-- OnboardingContainerFragment로의 목적지 정의 -->
-    <fragment
-        android:id="@+id/onboardingContainerFragment"
-        android:name="com.egobook.app.ui.onboarding.view.OnboardingContainerFragment"
-        android:label="온보딩"
-        tools:layout="@layout/fragment_onboarding_container">
-        <action
-            android:id="@+id/action_onboarding_to_home"
-            app:destination="@id/menu_home"
-            app:popUpTo="@id/onboardingContainerFragment"
-            app:popUpToInclusive="true" />
     </fragment>
 
     <fragment
